Added SOA and Eventbus and fixed General Joystick print

+++ src/main/java/Person/FLFOperator.java
@@ -10,5 +10,5 @@ public abstract class FLFOperator extends Person{
     public abstract void pressJoystickRight();
     public abstract void pressJoystickLeft();
     public abstract void pressFeeler();
-    public abstract void tiltJoystick(int[] axis);
+    public abstract void tiltJoystick(byte[] axis);
 }
